feat: update member-delimiter-style and comma-dangle
BREAKING CHANGE: comman-dangle is now `always-multiline` for `latest

This is change to make it aligns with `member-delimiter-style`.

Since the reason to use `only-multiline` is fore backward compatibility for es5 code,

it does not need to be applied in `latest` and in `TypeScript`.

So the rule is stricten in favor of a more uniform style.